Apricot Branch ==============
- Memory usage limit for undo, specified in megabytes, the default 0
  means unlimited.
- Multiple undo levels for image painting. Works separate from global
  undo similar to editmode undo, and undo for multiple images is
  supported.

+++ intern/memutil/intern/MEM_CacheLimiterC-Api.cpp
@@ -27,18 +27,18 @@
 #include "MEM_CacheLimiter.h"
 #include "MEM_CacheLimiterC-Api.h"
 
-static int & get_max() 
+static unsigned long & get_max() 
 {
-	static int m = 32*1024*1024;
+	static unsigned long m = 32*1024*1024;
 	return m;
 }
 
-void MEM_CacheLimiter_set_maximum(int m)
+void MEM_CacheLimiter_set_maximum(unsigned long m)
 {
 	get_max() = m;
 }
 
-int MEM_CacheLimiter_get_maximum()
+unsigned long MEM_CacheLimiter_get_maximum()
 {
 	return get_max();
 }
